Create basic ScriptureView and start ButtonView
I have a working view for Scriptureview and it is connected to the
Contentview.

I created the base file for ButtonView but have not put any code into
it yet as I need to plan how I am going to design the button.  Will use
onTap probably to trigger the action.

Still need to create the TabView.