Create placeholder graphic and JSON decoding
I created a very basic button to summon a random verse from the JSON.

I created a struct to hold Bible verses.

I added an extension to Bundle to decode JSON from the struc.  However,
the JSON I had created was causing fatal errors on the decode stage, so
I will have to go back and redo that.